+++ ProiectNoteElevi/src/main/java/note/repository/ClasaRepository.java
@@ -13,7 +13,7 @@ import note.model.Nota;
 public interface ClasaRepository {
 	
 	void creazaClasa(List<Elev> elevi, List<Nota> note);
-	HashMap<Elev, HashMap<String, List<Double>>> getClasa();
+	HashMap<Elev, HashMap<String, List<Integer>>> getClasa();
 	List<Medie> calculeazaMedii() throws ClasaException;
 	void afiseazaClasa();
 	List<Corigent> getCorigenti();
